refactor: 장바구니 뱃지 관련 코드 수정
