+++ app/src/main/res/layout/cart_action_layout.xml
@@ -46,7 +46,7 @@
             android:textColor="@android:color/white"
             android:textSize="12sp"
             android:gravity="center"
-            android:visibility="@{viewModel.isBadgeEnable? View.VISIBLE : View.GONE}" />
+            android:visibility="@{viewModel.cartCount > 0 ? View.VISIBLE : View.GONE}"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
